refactor: split hook message injector facade
