+++ src/features/hook-message-injector/sdk-message-lookup.ts
@@ -1,0 +1,122 @@
+import type { PluginInput } from "@opencode-ai/plugin"
+import { isCompactionMessage } from "../../shared/compaction-marker"
+import { log } from "../../shared/logger"
+import { normalizeSDKResponse } from "../../shared/normalize-sdk-response"
+import type { StoredMessage, ToolPermission } from "./types"
+
+export type OpencodeClient = PluginInput["client"]
+
+interface SDKMessage {
+  id?: string
+  info?: {
+    agent?: string
+    model?: {
+      providerID?: string
+      modelID?: string
+      variant?: string
+    }
+    providerID?: string
+    modelID?: string
+    tools?: Record<string, ToolPermission>
+    time?: {
+      created?: number
+    }
+  }
+  parts?: Array<{ type?: string }>
+}
+
+function convertSDKMessageToStoredMessage(msg: SDKMessage): StoredMessage | null {
+  if (isCompactionMessage(msg)) {
+    return null
+  }
+
+  const info = msg.info
+  if (!info) return null
+
+  const providerID = info.model?.providerID ?? info.providerID
+  const modelID = info.model?.modelID ?? info.modelID
+  const variant = info.model?.variant
+
+  if (!info.agent && !providerID && !modelID) {
+    return null
+  }
+
+  return {
+    agent: info.agent,
+    model: providerID && modelID
+      ? { providerID, modelID, ...(variant ? { variant } : {}) }
+      : undefined,
+    tools: info.tools,
+  }
+}
+
+export async function findNearestMessageWithFieldsFromSDK(
+  client: OpencodeClient,
+  sessionID: string
+): Promise<StoredMessage | null> {
+  try {
+    const response = await client.session.messages({ path: { id: sessionID } })
+    const emptyMessages: SDKMessage[] = []
+    const messages = normalizeSDKResponse(response, emptyMessages, { preferResponseOnMissingData: true })
+      .map((message) => ({
+        stored: convertSDKMessageToStoredMessage(message),
+        createdAt: message.info?.time?.created ?? Number.NEGATIVE_INFINITY,
+        id: typeof message.id === "string" ? message.id : "",
+      }))
+      .sort((left, right) => right.createdAt - left.createdAt || right.id.localeCompare(left.id))
+
+    for (const message of messages) {
+      const stored = message.stored
+      if (stored?.agent && stored.model?.providerID && stored.model?.modelID) {
+        return stored
+      }
+    }
+
+    for (const message of messages) {
+      const stored = message.stored
+      if (stored?.agent || (stored?.model?.providerID && stored?.model?.modelID)) {
+        return stored
+      }
+    }
+  } catch (error) {
+    const errorMessage = error instanceof Error ? error.message : String(error)
+    log("[hook-message-injector] SDK message fetch failed", {
+      sessionID,
+      error: errorMessage,
+    })
+  }
+  return null
+}
+
+export async function findFirstMessageWithAgentFromSDK(
+  client: OpencodeClient,
+  sessionID: string
+): Promise<string | null> {
+  try {
+    const response = await client.session.messages({ path: { id: sessionID } })
+    const emptyMessages: SDKMessage[] = []
+    const messages = normalizeSDKResponse(response, emptyMessages, { preferResponseOnMissingData: true })
+      .sort((left, right) => {
+        const leftTime = left.info?.time?.created ?? Number.POSITIVE_INFINITY
+        const rightTime = right.info?.time?.created ?? Number.POSITIVE_INFINITY
+        if (leftTime !== rightTime) return leftTime - rightTime
+        const leftId = typeof left.id === "string" ? left.id : ""
+        const rightId = typeof right.id === "string" ? right.id : ""
+        return leftId.localeCompare(rightId)
+      })
+
+    for (const msg of messages) {
+      const stored = convertSDKMessageToStoredMessage(msg)
+      if (stored?.agent) {
+        return stored.agent
+      }
+    }
+  } catch (error) {
+    const errorMessage = error instanceof Error ? error.message : String(error)
+    log("[hook-message-injector] SDK agent fetch failed", {
+      sessionID,
+      error: errorMessage,
+    })
+  }
+  return null
+}
